pyproject.toml
[tool.poetry]
name = "36-kuda-ugodno-backend-api"
version = "0.0.0"
description = "API сервис для веб-сайта 'Куда Угодно'"
authors = ["Антон Громов", "Никита", "Анна Боссир", "Татьяна Харламова"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.1.6"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.52"
celery = "^5.4.0"
redis = "^5.2.1"
django-celery-beat = "^2.7.0"
drf-spectacular = "^0.28.0"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"
dj-rest-auth = "^7.0.1"
django-allauth = "^65.4.0"
social-auth-app-django = "^5.4.2"
djangorestframework-simplejwt = "^5.4.0"
pytest = "^8.3.4"
pytest-django = "^4.10.0"
factory-boy = "^3.3.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = "migrations/"

[tool.isort]
include_trailing_comma = true
line_length = 119
lines_after_imports = 2
multi_line_output = 3
skip = ["__init__.py"]

[tool.flake8]
max-line-length = 119
exclude = ".git,__pycache__,poetry.lock,pyproject.toml,migrations,tests.py"